server/requirements.txt
fastapi>=0.116.1,<0.117.0
uvicorn>=0.35.0,<0.36.0
pydantic>=2.11.7,<3.0.0
langgraph>=0.6.6
langchain-google-genai>=2.1.9
python-dotenv>=1.1.1
psycopg2-binary>=2.9.10,<3.0.0
sqlalchemy>=2.0.43,<3.0.0
pydantic-settings>=2.10.1,<3.0.0
composio-core>=0.5.0,<1.0.0
alembic>=1.13.1,<2.0.0
composio-client>=0.5.0,<1.0.0
httpx>=0.24.0
aiofiles>=23.0.0
email-validator>=2.0.0
google-generativeai>=0.3.0